refactor: indexed the file name where it's needed
