--- src/features/groups/components/CreateGroupButton/index.ts
@@ -1,1 +1,0 @@
-export { CreateGroupButton } from './CreateGroupButton';
