--- src/features/groups/components/DeleteGroupsButton/index.ts
@@ -1,1 +1,0 @@
-export { DeleteGroupsButton } from './DeleteGroupsButton';
