--- src/features/groups/components/GroupCardList/index.ts
@@ -1,1 +1,0 @@
-export { GroupCardList } from './GroupCardList';
