--- src/features/groups/components/GroupCollectionMenuButton/index.ts
@@ -1,1 +1,0 @@
-export { GroupCollectionMenuButton } from './GroupCollectionMenuButton';
