--- src/features/groups/components/RenameGroupsButton/index.ts
@@ -1,1 +1,0 @@
-export { RenameGroupsButton } from './RenameGroupsButton';
